Import exact numbers from JANI.
This way we don't risk decimal->double conversion loss of precision.

It's almost certainly way overkill, but it's necessary for the
correctness of the --exact mode of DFTCalc.

On the bright side, it's not really a performance loss since we turn
everything into strings after parsing anyway.

+++ src/nl/utwente/ewi/fmt/EXPRES/Composition.java
@@ -1321,7 +1321,7 @@ public class Composition implements MarkableLTS
 				for (Object o : bound.keySet()) {
 					if ("upper".equals(o)) {
 						o = bound.get("upper");
-						timeBound = JaniUtils.getConstantDouble(o, constants);
+						timeBound = JaniUtils.getConstantDouble(o, constants).doubleValue();
 					} else if (!"upper-exclusive".equals(o)) {
 						throw new UnsupportedOperationException("Only constant-valued upper bounds currently supported.");
 					}
